requirements.txt
faiss-cpu
python-dotenv
pypdf
streamlit
google-genai
numpy
